legacy/python-backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain-google-genai>=2.0.0
langchain-groq
python-dotenv